backend/requirements.txt
fastapi
uvicorn
gunicorn
google-generativeai
pydantic
python-dotenv